requirements.txt
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.3.2
pretrainedmodels==0.7.4
scikit-learn==0.24.2
scipy==1.5.4
segmentation-models-pytorch==0.2.1
sklearn==0.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3